+++ wtfix/apps/admin.py
@@ -196,7 +196,8 @@ class HeartbeatApp(MessageTypeHandlerApp):
                 f"{self.name}: Unhandled exception while monitoring heartbeat! Shutting down pipeline..."
             )
             asyncio.create_task(self.pipeline.stop())
-            raise
+            # NOTE: We don't re-raise the exception here as this will cause it to interrupt stopping this app
+            # when we `await cancel_task` in stop().
 
     async def send_test_request(self):
         """