+++ wtfix/apps/brokers.py
@@ -54,6 +54,9 @@ class RedisPubSubApp(BaseApp):
                         self.send(message)
                     )  # Pass message on to pipeline
 
+        except asyncio.exceptions.CancelledError:
+            logger.info(f"{self.name}: {asyncio.current_task().get_name()} cancelled!")
+
         except aioredis.ChannelClosedError:
             # Shutting down...
             logger.info(f"{self.name}: Unsubscribed from {send_channel.name}.")
@@ -73,13 +76,7 @@ class RedisPubSubApp(BaseApp):
     async def stop(self, *args, **kwargs):
         if self._channel_reader_task is not None:
             self._channel_reader_task.cancel()
-            try:
-                await self._channel_reader_task
-            except asyncio.exceptions.CancelledError:
-                # Cancellation request received - close connections....
-                logger.info(
-                    f"{self.name}: {self._channel_reader_task.get_name()} cancelled!"
-                )
+            await self._channel_reader_task
 
         with await self.redis_pool as conn:
             await conn.unsubscribe(self.SEND_CHANNEL)